Re-organize app files
Make a separate file (resources/gunicorn_app.py) for the gunicorn class
app wrapper and move it into resources.
Make a separate file (app.py) for the construction of the application
(endpoints, resources, etc.).
Make a file (run_app.py) which imports the app defined in (app.py),
then convert it into a gunicorn application and launch it.